+++ src/TeamCityMcpTools/GlobalUsings.cs
@@ -1,6 +1,7 @@
 global using FluentResults;
 global using ModelContextProtocol.Server;
 global using System.ComponentModel;
+global using System.Linq;
 global using Microsoft.Extensions.DependencyInjection;
 global using Microsoft.Extensions.Logging;
 global using System.Text;